requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
pandas
openpyxl
jq